extracted getToken into http api
